Add release and environment tracking to Sentry
Release is the SHA of the current version of the code. They say it is
global across the organization, so it should be prefixed with something
project-specific (like locuszoom-api), but in practice I don't see that
happening in the Sentry UI.

Environment is just production, staging/dev, etc. You can filter by
environment in the very upper-right hand corner of the UI.